test: updates for running jest unit test in core
