+++ packages/salesforcedx-vscode-core/src/commands/util/conflictDetectionMessages.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2022, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export interface ConflictDetectionMessages {
+  warningMessageKey: string;
+  commandHint: (input: string | string[]) => string;
+}
