+++ scripts/setup-jest.ts
@@ -36,6 +36,9 @@ const getMockVSCode = () => {
     },
     EventEmitter: EventEmitter,
     ExtensionMode: { Production: 1, Development: 2, Test: 3 },
+    languages: {
+      createDiagnosticCollection: jest.fn()
+    },
     Uri: {
       parse: jest.fn(),
       file: jest.fn()
@@ -64,7 +67,11 @@ const getMockVSCode = () => {
         };
       },
       onDidChangeConfiguration: jest.fn(),
-      createFileSystemWatcher: jest.fn(),
+      createFileSystemWatcher: jest.fn().mockReturnValue({
+        onDidChange: jest.fn(),
+        onDidCreate: jest.fn(),
+        onDidDelete: jest.fn()
+      }),
       workspaceFolders: []
     }
   };
